requirements.txt
pywebview==5.1
fastapi==0.104.1
uvicorn==0.24.0
pyinstaller==6.4.0
